requirements.txt
google-api-python-client==2.33.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
pandas==1.3.5
camelot-py[cv]==0.10.1